requirements.txt
aiohappyeyeballs==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.10.5 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
dataclasses-json==0.6.7 ; python_version >= "3.12" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
flask==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-auth-oauthlib==1.2.1 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.34.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-firestore==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.0.3 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
grpcio-status==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.12" and python_version < "4.0"
idna==3.8 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
jiter==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.12" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
langchain-community==0.2.16 ; python_version >= "3.12" and python_version < "4.0"
langchain-core==0.2.38 ; python_version >= "3.12" and python_version < "4.0"
langchain-google-firestore==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
langchain-openai==0.1.23 ; python_version >= "3.12" and python_version < "4.0"
langchain-text-splitters==0.2.4 ; python_version >= "3.12" and python_version < "4.0"
langchain==0.2.16 ; python_version >= "3.12" and python_version < "4.0"
langsmith==0.1.116 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
marshmallow==3.22.0 ; python_version >= "3.12" and python_version < "4.0"
more-itertools==10.5.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.12" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.12" and python_version < "4.0"
openai==1.44.0 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.7 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.23.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
regex==2024.7.24 ; python_version >= "3.12" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.34 ; python_version >= "3.12" and python_version < "4.0"
tenacity==8.5.0 ; python_version >= "3.12" and python_version < "4.0"
tiktoken==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.5 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
werkzeug==3.0.4 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.10.0 ; python_version >= "3.12" and python_version < "4.0"
